Move docker command test
